pyproject.toml
[project]
name = "streamdeck-notify"
version = "0.1.0"
description = "Stream Deck as a notification center"
requires-python = ">=3.11"
dependencies = [
    "streamdeck>=0.9.5",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.1",
    "dbus-next>=0.2.3",
    "cairosvg>=2.7",
]

[project.optional-dependencies]
dev = ["ruff", "mypy"]

[project.scripts]
streamdeck-notify = "src.daemon:main"
